Add support for the new Beaver config option.
Beaver >= 31 now requires you to set the config option ```logstash_version```. This has been added to support the new Logstash 1.2 JSON schema. Assuming that the same cookbook is being used elsewhere in the infrastructure to install Logstash, this small change checks for the version of Logstash and adds the appropriate value for ```logstash_version``` to the ```conf``` hash.